+++ CHANGES.md
@@ -1,4 +1,7 @@
 ## ?
+### Added
+- Add `Conformist.list` to support decoding list of `'a`
+
 ### Changed
 - `Conformist.Field.optional` is now `Conformist.Field.is_optional` to avoid confusion with `Conformist.optional`
 - The encoded value is assumed to be `string list` instead of `string`. This is preparation to support decoding of a list of strings.